+++ CHANGE_LOG.md
@@ -1,3 +1,9 @@
+## Version 0.7.0
+* Removed `RubyInLine`
+* Removed possibility to return false from `irr_guess`
+* Removed global newton module
+* Bumped dependencies
+
 ## Version 0.5.4
 * Fallsback If Newton Methods returns NaN